Refactored profile
